+++ R/cloudUploaders.R
@@ -2,12 +2,12 @@
 
 #' Backup R package list
 #'
-#' @param backupPath The path to where \code{\link{savePkgs}} should write the
+#' @param backupPath The path to where \code{\link{save_pkgs}} should write the
 #' list of packages, optional
 #' @param filename The name of the csv file to save the package names, optional
 #' @param cloudProvider Optional, if specified the package will attempt to write
 #' the package list to a tempfile and then upload that tempfile to a cloud service
-#' @param ... Additional arguments to pass to \code{\link{writeCloud}}
+#' @param ... Additional arguments to pass to \code{\link{write_cloud}}
 #' @return Nothing
 #' @export
 #'
@@ -20,12 +20,12 @@ pkg_backup <- function(backupPath = NULL, filename = NULL,
     filename <- paste0("RPackageBackup_", Sys.Date(), ".csv")
     fullFN <- file.path(backupPath, filename)
   }
-  savePkgs(backupPath, filename=filename)
+  save_pkgs(backupPath, filename=filename)
   if(missing(cloudProvider)){
     message("List of R packages backed up successfully")
     message(paste0("Package list saved at ", paste(backupPath, filename, sep = "/")))
   } else{
-    writeCloud(filename = fullFN, provider = cloudProvider, ...)
+    write_cloud(filename = fullFN, provider = cloudProvider, ...)
     message(paste0("List of R packages backed up to ", cloudProvider))
   }
 }
@@ -33,7 +33,7 @@ pkg_backup <- function(backupPath = NULL, filename = NULL,
 #' Restore R packages from a package list stored somewhere else
 #'
 #' @param install logical, should packages in the package list be installed?
-#' @param backupPath The path to where \code{\link{savePkgs}} should write the
+#' @param backupPath The path to where \code{\link{save_pkgs}} should write the
 #' list of packages, optional
 #' @param filename The name of the csv file to save the package names, optional
 #' @param cloudProvider Optional, if specified the package will attempt to write
@@ -41,7 +41,7 @@ pkg_backup <- function(backupPath = NULL, filename = NULL,
 #' @param libpath Path to the library
 #' @param update A logical, should packages be updated first? Default is false
 #' @param keep_all A logical, should all packages in pkgList not local be installed?
-#' @param ... Additional arguments to pass to \code{\link{readCloud}}
+#' @param ... Additional arguments to pass to \code{\link{read_cloud}}
 #'
 #' @return Nothing.
 #' @export
@@ -49,13 +49,13 @@ pkg_restore <- function(install = TRUE, backupPath = NULL, filename = NULL,
                         cloudProvider = NULL, libpath = NULL, update = FALSE,
                         keep_all = FALSE, ...){
   if(!missing(cloudProvider)){
-    pkgList <- readCloud(filename = filename, provider = cloudProvider, ...)
+    pkgList <- read_cloud(filename = filename, provider = cloudProvider, ...)
   } else {
-    pkgList <- readPkgs(backupPath, filename=filename)
+    pkgList <- read_pkgs(backupPath, filename=filename)
   }
   if(install){
     # First validate
-    installPkgs(pkgList, libpath = libpath, update = update,
+    install_pkgs(pkgList, libpath = libpath, update = update,
                 keep_all = keep_all)
   } else{
     pkgList <- sync_pkgs(pkgList = pkgList, keep_all = keep_all)
@@ -78,7 +78,7 @@ pkg_restore <- function(install = TRUE, backupPath = NULL, filename = NULL,
 #' @return Nothing.
 #' @seealso \code{\link[rdrop2:drop_read_csv]{drop_upload}}
 #' @export
-writeCloud <- function(filename, provider = c("dropbox"), ...){
+write_cloud <- function(filename, provider = c("dropbox"), ...){
   provider <- match.arg(provider,
                         choices = c("dropbox"), several.ok = FALSE)
   if(provider == "dropbox"){
@@ -97,7 +97,7 @@ writeCloud <- function(filename, provider = c("dropbox"), ...){
 #' @return Nothing.
 #' @seealso \code{\link[rdrop2:drop_read_csv]{drop_upload}}
 #' @export
-readCloud <- function(filename, provider = c("dropbox"), ...){
+read_cloud <- function(filename, provider = c("dropbox"), ...){
   provider <- match.arg(provider,
                         choices = c("dropbox"), several.ok = FALSE)
   if(provider == "dropbox"){