+++ NEWS.md
@@ -1,5 +1,6 @@
 # maintainr 0.1
 
+* standardize function names to snake_case
 * Modernize package to include latest devtools features
 * Initial push